add bootcamp and courses ownership

+++ controllers/courses.js
@@ -36,9 +36,19 @@ exports.getSingleCourse = asyncHandler(async (req, res, next) => {
 //api/v1/bootcamps/:bootcampId/courses
 exports.createCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
   if (!bootcamp) {
     return next(new ErrorResponse("Bootcamp not found", 404));
+  }
+  //Make sure is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized for add course to bootcamp ${bootcamp._id}`,
+        401
+      )
+    );
   }
   const course = await Course.create(req.body);
   res.json({ success: true, data: course });
@@ -47,24 +57,42 @@ exports.createCourse = asyncHandler(async (req, res, next) => {
 //private
 //api/v1/courses/:id
 exports.updateCourse = asyncHandler(async (req, res, next) => {
-  const course = await Course.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true
-  });
+  let course = await Course.findById(req.params.id);
   if (!course) {
     return next(new ErrorResponse("Course not found", 404));
   }
 
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized for update course ${course._id}`,
+        401
+      )
+    );
+  }
+  course = await Course.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true
+  });
   res.json({ success: true, data: course });
 });
 
 //private
 //api/v1/courses/:id
 exports.deleteCourse = asyncHandler(async (req, res, next) => {
-  const course = await Course.findByIdAndDelete(req.params.id);
+  const course = await Course.findById(req.params.id);
   if (!course) {
     return next(new ErrorResponse("Course not found", 404));
   }
 
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized for delete course ${course._id}`,
+        401
+      )
+    );
+  }
+  await Course.findByIdAndDelete(req.params.id);
   res.json({ success: true });
 });
